rebased on master, incorporated review comments
