+++ packages/angular-material/package.json
@@ -49,7 +49,6 @@
     "@jsonforms/core": "^2.0.0-rc.2",
     "core-js": "^2.5.3",
     "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0",
     "rxjs": "^5.5.6",
     "zone.js": "^0.8.20"
   },
