+++ packages/angular/src/index.ts
@@ -1,4 +1,3 @@
 export * from './base.renderer';
 export * from './jsonforms.component';
 export * from './jsonforms.module';
-// export * from './util';
