+++ packages/angular/src/unknown.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 @Component({
     selector: 'unknown.renderer',
-    template: 'Unknown Renderer!'
+    template: 'No applicable renderer found!'
 })
 export class UnknownRenderer {
 }
