--- packages/core/src/renderers/index.ts
@@ -1,5 +1,0 @@
-export * from './Control';
-export * from './JsonForms';
-export * from './DispatchField';
-export * from './Renderer';
-export * from './Formatted';
