+++ packages/core/src/util/index.ts
@@ -56,3 +56,4 @@ export { isEnabled, isVisible, Runtime };
 export * from './renderer';
 export * from './field';
 export * from './runtime';
+export * from './Formatted';
