+++ packages/react/src/JsonForms.tsx
@@ -2,9 +2,10 @@ import * as React from 'react';
 import * as _ from 'lodash';
 import { connect } from 'react-redux';
 import { UnknownRenderer } from './UnknownRenderer';
-import { JsonFormsProps, mapStateToDispatchRendererProps } from '@jsonforms/core';
+import { DispatchRendererProps, mapStateToDispatchRendererProps } from '@jsonforms/core';
 
-const JsonFormsDispatchRenderer = ({ uischema, schema, path, renderers }: JsonFormsProps) => {
+const JsonFormsDispatchRenderer =
+  ({ uischema, schema, path, renderers }: DispatchRendererProps) => {
   const renderer = _.maxBy(renderers, r => r.tester(uischema, schema));
   if (renderer === undefined || renderer.tester(uischema, schema) === -1) {
     return <UnknownRenderer type={'renderer'}/>;
